+++ Changelog.md
@@ -1,7 +1,14 @@
 # Changelog
 
+## 2.0.0
+- Added option to only skip certain number of levels at every use
+- Added configurable list of levels not to be skipped in mine or skull cavern, defaults to levels 100, 200 and 300 in skull cavern
+
+## 1.0.2
+- Added configuration to not skip levels with certain number of certain monsters
+
 ## 1.0.1
-- Number of used staircases is now properly increased when using staircases in bulk
+- Number of used staircases is now properly decreased when using staircases in bulk
 - Optimized code in case all levels are skipped to prevent unnecessary loading of intermediate levels
 - Added option for (not-) skipping mushroom levels